backend/requirements.txt
# Agent Nexus Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# File watching
watchdog>=3.0.0

# Database
aiosqlite>=0.19.0

# HTTP client for API
httpx>=0.26.0

# Security
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
